requirements.txt
streamlit
Flask==3.1.3
tensorflow==2.15.0
numpy
pillow
scikit-learn
pandas